requirements.txt
asgiref==3.8.1
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==23.0.0
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
uritemplate==4.1.1
